# fix(expr-common): Coerce to Decimal(20, 0) when combining UInt64 with signed integers