Previously, when combining `UInt64` with any signed integer, the resulting type would be `Int64`, which would result in lost information. Now, combining `UInt64` with a signed integer results in a `Decimal(20, 0)`, which is able to encode all (64-bit) integer types. Thanks @jonahgao for the pointers.

The function `bitwise_coercion` remains the same, since it's probably not a good idea to introduce decimals when performing bitwise operations. In this case, it converts `(UInt64 | _)` to `UInt64`.

## Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and enhancements and this helps us generate change logs for our releases. You can link an issue to this PR using the GitHub syntax. For example `Closes #123` indicates that this PR will close issue #123.
-->

Closes #14208.

## What changes are included in this PR?

<!--
There is no need to duplicate the description in the issue here but it is sometimes worth providing a summary of the individual changes in this PR.
-->

- Updated `binary_numeric_coercion` in `expr-common/type_coercion/binary.rs`.
- Added new tests to `expr-common/type_coercion/binary.rs`.
- Updated existing sqllogic tests to use the new coercion.

## Are these changes tested?

<!--
We typically require tests for all PRs in order to:
1. Prevent the code from being accidentally broken by subsequent changes
2. Serve as another way to document the expected behavior of the code

If tests are not included in your PR, please explain why (for example, are they covered by existing tests)?
-->

Yes.

## Are there any user-facing changes?

<!--
If there are user-facing changes then we may require documentation to be updated before approving the PR.
-->

<!--
If there are any breaking changes to public APIs, please add the `api change` label.
-->

No.