@jonahgao 2025-01-21
Perhaps we should add a sqllogictest test for #14208.