@nuno-faria 2025-01-21
> Perhaps we should add a sqllogictest test for #14208.

Done.