@ozankabak 2025-01-24
When we combine a `UInt64` with a literal value (say `5`), we don't accidentally consider the literal value a signed value and do this coercion to decimal, right? That would kill performance in many cases.