@ozankabak 2025-01-24
This is a problem with a large potential performance impact in many use cases. I do not think it is a good idea to defer its solution with an open-ended timeframe. I see two ways to proceed:

1. If it is relatively easy, we can do a quick follow-on PR to parse/type literals more intelligently. Maybe always parse as the smallest/narrowest datatype that will hold the constant. Coercion will then widen the type if necessary.
2. If doing (1) is not easy, we should revert this PR to avoid the performance impact and think of a more holistic solution that includes a mechanism for doing something like (1), or a more involved mechanism utilizing statistics.

I am fine with both ways forward.