@alamb 2025-01-24
@ozankabak note the coercion rules @jonahgao refers to above in https://github.com/apache/datafusion/pull/14223#issuecomment-2611707619 are for `UNION`, not for example, integer arithmetic)

I am writing some tests to show what is happening