@alamb 2025-01-24
Here is a PR that tries to demonstrate more of what is happening:
- https://github.com/apache/datafusion/pull/14270