@ozankabak 2025-01-24
Comparison with a literal is a very common pattern. I am still worried.