@berkaysynnada 2025-01-24
I think this change brings more harm than it helps. I wrote a simple benchmark:

* Table having 100_000 rows, and one column in UInt64
* The values are randomly generated between 0 and 100_000_000
* The query is `SELECT a > -1 FROM random_a`

**after this commit:**
 ```
SELECT a > -1
                        time:   [1.3010 ms 1.4096 ms 1.5468 ms]
Found 13 outliers among 100 measurements (13.00%)
  6 (6.00%) high mild
  7 (7.00%) high severe
```
  
**before this commit:**
```
SELECT a > -1
                        time:   [601.85 µs 604.89 µs 610.50 µs]
                        change: [-60.997% -57.120% -53.557%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  1 (1.00%) low mild
  1 (1.00%) high mild
  3 (3.00%) high severe
  ```