@jonahgao 2025-01-24
> Comparison with a literal is a very common pattern. I am still worried.

I found that `unwrap_cast_in_comparison` can optimize this case.
```sh
DataFusion CLI v44.0.0
create table t(a bigint unsigned);

> explain verbose select a > 1 from t;
| logical_plan after simplify_expressions                    | Projection: CAST(t.a AS Decimal128(20, 0)) > Decimal128(Some(1),20,0) AS t.a > Int64(1)                                |
|                                                            |   TableScan: t                                                                                                         |
| logical_plan after unwrap_cast_in_comparison               | Projection: t.a > UInt64(1) AS t.a > Int64(1)                                                                          |
|                                                            |   TableScan: t                                                                                                         |
```