@berkaysynnada 2025-01-24
> > I think this change brings more harm than it helps. I wrote a simple benchmark:
> > Table having 100_000 rows, and one column in UInt64
> > The values are randomly generated between 0 and 100_000_000
> 
> The problem is that if the column contains values greater than 9223372036854775807(`i64::MAX`), this query can not run.
> 
> I think this is a necessary overhead for correctness, unless we can guarantee that the column's value will not exceed `i64::MAX`(through statistics), or arrow-rs supports comparing two different types.

If the user believes such a scenario is possible, they should be expected to explicitly provide the casted type. Why do we impose this overhead on all cases?