@jonahgao 2025-01-24
> If the user believes such a scenario is possible, they should be expected to explicitly provide the casted type. Why do we impose this overhead on all cases?

The user may know nothing about the data in the table and be unable to make decisions. Additionally,  some query statements are generated at runtime. I think this is the optimizer's responsibility, for example, to introduce a physical optimization rule similar to `unwrap_cast_in_comparison`.